+++ packages/devtools_app/lib/src/debugger/codeview.dart
@@ -24,7 +24,6 @@ import '../ui/colors.dart';
 import '../ui/search.dart';
 import '../ui/utils.dart';
 import '../utils.dart';
-import '../vm_service_utils.dart';
 import 'breakpoints.dart';
 import 'common.dart';
 import 'debugger_controller.dart';
@@ -417,53 +416,7 @@ class _CodeViewState extends State<CodeView>
         ),
       ),
     );
-  }
-
-  Widget buildCodeviewTitle(ThemeData theme) {
-    return ValueListenableBuilder(
-      valueListenable: widget.controller.scriptsHistory.current,
-      builder: (context, scriptsHistory, _) {
-        return debuggerSectionTitle(
-          theme,
-          child: Row(
-            children: [
-              ToolbarAction(
-                icon: Icons.chevron_left,
-                onPressed:
-                    scriptsHistory.hasPrevious ? scriptsHistory.moveBack : null,
-              ),
-              ToolbarAction(
-                icon: Icons.chevron_right,
-                onPressed:
-                    scriptsHistory.hasNext ? scriptsHistory.moveForward : null,
-              ),
-              const SizedBox(width: denseSpacing),
-              const VerticalDivider(thickness: 1.0),
-              const SizedBox(width: defaultSpacing),
-              Expanded(
-                child: Text(
-                  scriptRef?.uri ?? ' ',
-                  style: theme.textTheme.subtitle2,
-                ),
-              ),
-              const SizedBox(width: denseSpacing),
-              ScriptPopupMenu(widget.controller),
-              const SizedBox(width: denseSpacing),
-              ScriptHistoryPopupMenu(
-                itemBuilder: _buildScriptMenuFromHistory,
-                onSelected: (scriptRef) {
-                  widget.controller
-                      .showScriptLocation(ScriptLocation(scriptRef));
-                },
-                enabled: scriptsHistory.hasScripts,
-              ),
-              const SizedBox(width: denseSpacing),
-            ],
-          ),
-        );
-      },
-    );
-  }
+  } 
 
   List<PopupMenuEntry<ScriptRef>> _buildScriptMenuFromHistory(
     BuildContext context,
