+++ packages/devtools_app/lib/src/inspector/diagnostics_node.dart
@@ -11,10 +11,10 @@ import 'package:flutter/material.dart';
 import 'package:flutter/rendering.dart';
 import 'package:vm_service/vm_service.dart';
 
+import '../debugger/debugger_model.dart';
 import '../enum_utils.dart';
 import '../ui/icons.dart';
 import '../utils.dart';
-import '../vm_service_utils.dart';
 import 'inspector_service.dart';
 
 final diagnosticLevelUtils = EnumUtils<DiagnosticLevel>(DiagnosticLevel.values);
