+++ packages/devtools_app/pubspec.yaml
@@ -129,5 +129,3 @@ dependency_overrides:
     path: ../devtools_shared #OVERRIDE_FOR_DEVELOPMENT
   devtools:  #OVERRIDE_FOR_DEVELOPMENT
     path: ../devtools  #OVERRIDE_FOR_DEVELOPMENT
-  vm_service:
-    path: ../../../sdk/pkg/vm_service
